requirements.txt
# Core
django==4.1.12
python-dotenv==1.0.0

# UI 
django-admin-soft-dashboard==1.0.12
# django-bootstrap-datepicker-plus==5.0.4

# Deployment
whitenoise==6.5.0
gunicorn==21.2.0

psycopg2-binary
# mysqlclient

# Logic
github3.py==4.0.1
django_extensions==3.2.3
oauthlib==3.2.2